change intger into decimal

+++ python-functions/2-temperature.py
@@ -1,3 +1,3 @@
 def convert_to_celsius(fahrenheit):
-    Celsius = int(round(fahrenheit - 32) * (5/9))
+    Celsius = float(round(fahrenheit - 32) * (5/9))
     return Celsius
